+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.7.1-SNAPSHOT]
 
+- paddingLeft/paddingRight converted from Double to Number
 - JsonObject.contains(text) and JsonModel.contains(text)
 - Fixed
 - Button.action() shortcut istead of Button.setOnAction()